requirements.txt
amqp==5.3.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.1
bleach==6.1.0
boto3==1.34.66
botocore==1.34.66
celery==5.5.0
certifi==2024.2.2
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
distro==1.9.0
Django==5.0.3
django-appconf==1.0.6
django-compat==1.0.15
django-cors-headers==4.3.1
django-filter==23.5
django-redis==5.4.0
django-select2==8.1.2
django-storages==1.14.2
django-summernote==0.8.20.0
djangorestframework==3.14.0
drf-yasg==1.21.8
Faker==37.1.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
icalendar==5.0.12
idna==3.6
inflection==0.5.1
jmespath==1.0.1
kombu==5.5.2
msgpack==1.0.8
openai==1.23.6
packaging==24.0
pillow==10.3.0
prompt_toolkit==3.0.50
psycopg==3.2.4
psycopg2==2.9.9
pydantic==2.7.1
pydantic_core==2.18.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
redis==5.0.2
requests==2.31.0
requests-toolbelt==1.0.0
s3transfer==0.10.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
